# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type InventoryItemSpecifications = { [key: string]: unknown };

export interface InventoryItem {
  id: string;
  name: string;
  sku: string;
  description?: string;
  category?: string;
  specifications?: InventoryItemSpecifications;
  currentStock: number;
  minimumStock?: number;
  maximumStock?: number;
  unit: string;
  unitCost: number;
  location?: string;
  supplier?: string;
  leadTime?: number;
  lastUpdated?: string;
  createdAt?: string;
}

export type InventoryItemInputSpecifications = { [key: string]: unknown };

export interface InventoryItemInput {
  name: string;
  sku: string;
  description?: string;
  category?: string;
  specifications?: InventoryItemInputSpecifications;
  currentStock: number;
  minimumStock?: number;
  maximumStock?: number;
  unit: string;
  unitCost: number;
  location?: string;
  supplier?: string;
  leadTime?: number;
}