# lib/api-zod/src/generated/types/inventoryItem.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryItemSpecifications } from './inventoryItemSpecifications';

export interface InventoryItem {
  id: string;
  name: string;
  sku: string;
  description?: string;
  category?: string;
  specifications?: InventoryItemSpecifications;
  currentStock: number;
  minimumStock?: number;
  maximumStock?: number;
  unit: string;
  unitCost: number;
  location?: string;
  supplier?: string;
  leadTime?: number;
  lastUpdated?: Date;
  createdAt?: Date;
}