# lib/api-zod/src/generated/types/inventoryItemInput.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryItemInputSpecifications } from './inventoryItemInputSpecifications';

export interface InventoryItemInput {
  name: string;
  sku: string;
  description?: string;
  category?: string;
  specifications?: InventoryItemInputSpecifications;
  currentStock: number;
  minimumStock?: number;
  maximumStock?: number;
  unit: string;
  unitCost: number;
  location?: string;
  supplier?: string;
  leadTime?: number;
}